% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{concat_exprs}
\alias{concat_exprs}
\title{Concatenates a list of flow frames into a tibble}
\usage{
concat_exprs(flowframe_list)
}
\arguments{
\item{flowframe_list}{a named list of flow frames}
}
\value{
a tibble
}
\description{
Concatenates a list of flow frames into a tibble
}
