% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygon_operations.R
\name{is_point_in_path}
\alias{is_point_in_path}
\title{Check if a point is contained in a polygon}
\usage{
is_point_in_path(x, y, poly)
}
\arguments{
\item{x}{The x coordinates of point.}

\item{y}{The y coordinates of point.}

\item{poly}{a list of numeric vectors defining the polygon
list(c(x, y), c(x, y), ...) or a data.frame
data.frame(x = c(0,0,1), y = c(0,1,1))}
}
\value{

}
\description{
based on: https://en.wikipedia.org/wiki/Even%E2%80%93odd_rule
}
\examples{
poly_selection <- data.frame(
    Petal.Length = c(2,4,6),
    Sepal.Width = c(2,4,2))

dplyr::mutate(
    iris,
    point_in_poly = is_point_in_path(
        Petal.Length,
        Sepal.Width,
        poly_selection))
}
