% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read.multi.FCS}
\alias{read.multi.FCS}
\title{Read FCS files with multiple experiments in them}
\usage{
read.multi.FCS(filename, ...)
}
\arguments{
\item{filename}{name of the file to be read}

\item{...}{additional arguments passed to flowCore::read.FCS}
}
\value{
a list of flow frames and warning of
        which ones were not read successfully
}
\description{
Read FCS files with multiple experiments in them
}
