% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygon_operations.R
\name{scale_polygon}
\alias{scale_polygon}
\title{Scale polygons around their center}
\usage{
scale_polygon(poly, scaling)
}
\arguments{
\item{poly}{a data frame defining a polygon (x and y coordinates)}

\item{scaling}{a numeric vector that defines the scaling factor}
}
\value{
data.frame defining a scaled polygon
}
\description{
Scale polygons around their center
}
\examples{
scale_polygon(data.frame(x = c(1, 2, 3), y = c(1, 2, 1)), 2)

poly <- data.frame(x = c(1, 1, 0, 0), y = c(1, 0, 0, 1))
library(ggplot2)
ggplot(data = NULL, aes(x = x, y = y)) +
  geom_polygon(data = poly, fill = "black") +
  geom_polygon(data = scale_polygon(poly, 0.5), fill = "gray") +
  geom_polygon(data = scale_polygon(poly, 2), alpha = 0.2, fill = "red") +
  geom_polygon(data = scale_polygon(poly, c(1, 3)), alpha = 0.2, fill = "green") +
  geom_polygon(data = scale_polygon(poly, c(3, 1)), alpha = 0.2, fill = "blue")
}
